Refactor per-season game logic into SeasonGameConfig classes
Move all year-keyed if-else chains and dicts from helper files into
a new game_specific/ package:
- base.py: SeasonGameConfig base class with 10 overridable methods
- registry.py: get_game(year) factory returning per-season instance
- seasons/game_specifics_NNNN.py: one file per season (2006–2026)

Update helpers to delegate to get_game():
- match_tiebreakers.py: remove all _tiebreak_YYYY classmethods
- event_insights_helper.py: remove all calculate_event_insights_YYYY
- matchstats_helper.py: replace MANUAL_COMPONENTS dict
- prediction_helper.py: replace relevant_stats if-else chain
- playoff_advancement_helper.py: replace ROUND_ROBIN_* dicts
- rankings_helper.py: replace NO_RECORD_YEARS/QUAL_AVERAGE_YEARS sets
- score_breakdown_keys.py: delegate valid_score_breakdown_keys()
- web/handlers/event.py + admin/event.py: replace ROUND_ROBIN_TIEBREAKERS

Also include 2026 tiebreak logic and test data from upstream #9610.
