+++ src/backend/common/game_specific/seasons/game_specifics_2015.py
@@ -1,0 +1,34 @@
+from typing import List, Optional, Set
+
+from backend.common.game_specific.base import SeasonGameConfig
+from backend.common.models.ranking_sort_order_info import RankingSortOrderInfo
+
+
+class GameSpecifics2015(SeasonGameConfig):
+    def record_in_rankings(self) -> bool:
+        return False
+
+    def qual_average_in_rankings(self) -> bool:
+        return True
+
+    def ranking_sort_order_info(self) -> Optional[List[RankingSortOrderInfo]]:
+        return [
+            {"name": "Qual Avg.", "precision": 1},
+            {"name": "Coopertition", "precision": 0},
+            {"name": "Auto", "precision": 0},
+            {"name": "Container", "precision": 0},
+            {"name": "Tote", "precision": 0},
+            {"name": "Litter", "precision": 0},
+        ]
+
+    def valid_score_breakdown_keys(self) -> Set[str]:
+        return set(
+            [
+                "coopertition_points",
+                "auto_points",
+                "container_points",
+                "tote_points",
+                "litter_points",
+                "foul_points",
+            ]
+        )
